requirements.txt
asgiref==3.3.1
Django==3.1.5
django-admin==2.0.1
django-cors-headers==3.6.0
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-six==1.0.4
djangorestframework==3.12.2
mysqlclient==2.0.3
pytz==2020.5
screen==1.0.1
sqlparse==0.4.1
xlwt==1.3.0
